# swagger/swagger.yaml
definitions:
  dto.CheckClaimDTO:
    properties:
      is_long:
        description: |-
          IsLong controls FormattedMessage's verbosity: false (the default —
          note the zero-value already means short, no need to set this at all
          for the common case) shows a brief paraphrased explainer per
          sub-claim; true shows the full reasoning plus its evidence/citations.
          Only affects FormattedMessage — SubClaims/Citations are always fully
          populated in the structured response either way.
        type: boolean
      text:
        description: |-
          Text is the claim or message to check — may be a direct question, a
          casual claim, or a sentence pulled from a video transcript.
        type: string
    required:
    - text
    type: object
  dto.CitationResponse:
    properties:
      chunk_id:
        description: |-
          ChunkID is the chunk this evidence traces back to — combine with
          DocumentID to link directly to this evidence (e.g. the standalone
          /evidence PDF viewer: /evidence?document_id=...&chunk_id=...).
        type: string
      document_id:
        description: DocumentID is the source document this evidence came from.
        type: string
      index:
        description: |-
          Index is the same 0-based number the judge's Reasoning text references
          inline as "[N]" — use it to link a prose citation back to this entry.
        type: integer
      source:
        description: Source is "chunk" or "fact".
        type: string
      text:
        description: Text is the matched text — a chunk's text, or a fact's subject-predicate-object.
        type: string
    type: object
  dto.ClaimCheckResponse:
    properties:
      formatted_message:
        description: |-
          FormattedMessage is a human-facing, color-coded (emoji-per-verdict)
          rendering of OverallSummary and every sub-claim's verdict/reasoning —
          meant to be displayed as-is by a chat client. Plain text, no
          HTML/Markdown markup. Empty when InScope is false.
        type: string
      in_scope:
        description: |-
          InScope is false if the input wasn't a health-related inquiry at all —
          this system doesn't take inquiries unrelated to health.
        type: boolean
      overall_summary:
        description: |-
          OverallSummary is a short rollup across every sub-claim's verdict.
          Empty when InScope is false.
        type: string
      refusal_reason:
        description: RefusalReason explains why InScope is false. Empty when InScope
          is true.
        type: string
      sub_claims:
        description: SubClaims is the per-sub-claim breakdown. Empty when InScope
          is false.
        items:
          $ref: '#/definitions/dto.SubClaimResponse'
        type: array
    type: object
...
  dto.SubClaimResponse:
    properties:
      brief_reasoning:
        description: |-
          BriefReasoning is a short, self-contained paraphrase of Reasoning,
          safe to display without the citations list alongside it.
        type: string
      citations:
        description: |-
          Citations are the specific pieces of knowledge-base evidence the
          verdict is based on. Empty whenever Verdict is insufficient_evidence.
        items:
          $ref: '#/definitions/dto.CitationResponse'
        type: array
      claim:
        description: Claim is the normalized sub-claim text that was actually checked.
        type: string
      high_risk_caution:
        description: |-
          HighRiskCaution is true when the forced-RED safety override fired
          (response-rule.txt section 4: tobacco/alcohol/illicit drugs, an
          unregulated homemade mixture, or self-medication) — Verdict is forced
          to "contradicted" whenever this is true, regardless of what the
          evidence otherwise showed.
        type: boolean
      reasoning:
        description: Reasoning is the judge's explanation for the verdict.
        type: string
      recommend_medical_consult:
        description: |-
          RecommendMedicalConsult is true whenever the claim's subject matter
          warrants recommending a real doctor — this system provides quick
          access to reliable information, it does not replace one.
        type: boolean
      severity:
        description: 'Severity is one of: routine, serious, emergency.'
        type: string
      verdict:
        description: |-
          Verdict is one of: supported, contradicted, partially_supported, mixed, insufficient_evidence.
          insufficient_evidence never implies the claim itself is right or wrong
          — only that the knowledge base had nothing relevant to check it against.
        type: string
    type: object
...
  model.APIResponse:
    properties:
      data: {}
      error:
        type: string
      request_id:
        format: uuid
        type: string
      status:
        maximum: 599
        minimum: 100
        type: integer
    required:
    - request_id
    - status
    type: object
...
info:
  contact: {}
paths:
  /api/v1/claims/check:
    post:
      consumes:
      - application/json
      description: Checks a claim (e.g. from a video transcript) against the knowledge
        base and returns a cited verdict. Out-of-scope input (not health-related)
        is a normal 200 response with in_scope=false, not an error.
      parameters:
      - description: Check Claim Form
        in: body
        name: form
        required: true
        schema:
          $ref: '#/definitions/dto.CheckClaimDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ClaimCheckResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/model.APIResponse'
            - properties:
                error:
                  type: string
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/model.APIResponse'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/model.APIResponse'
            - properties:
                error:
                  type: string
              type: object
      summary: Check a claim
      tags:
      - claims
...
swagger: "2.0"